Added Basic Support for Uno and Codenames Decks
Progress:
* Uno and Codenames decks can now be generated.
* Added a "Schematics" type that is used to denote the numbers and
values of cards to be associated with a given card suit/type.
  - Deck Manifests now use schematics to determine the correct
arrangement of cards. This only works for static decks.
* Added a function to generate a Codenames Manifest. (Decks that are more
dynamic in nature will likely require more robust functions for
determining card arrangement.)
* "Codenames" has been added to "CardValues" as an array of strings.
This will very possibly be changed in the future since this is so simple
and raises the potential for accidental duplicate values.
* Added a function to generate card values from schematics instead of
from enumerations. The old function should be considered "deprecated",
at least for now.
* Added a function to obtain a random sample of size k from a population
of size n. The function allows for modifications to the original
population if needed, as in the case of Codenames.
* Updated notes with some useful links.

Concerns and Todo's:
* The comments in the random sampling function may be too large. It's
worth considering moving those details into NOTES.md.
* It's worth putting more thought into Manifests/Schematics. Is there
any way in which we are duplicating a representation? Maybe not?
  - On the same note, it's worth putting good thought into what
functions are used to generate manifests (or to help with the process).
* Codenames values as an array of strings is likely dangerous.
Considerations for a different representation are worth your time.