Add Basic functionality for a Codenames Game
Known Progress:
* Added an abstract base Game class for all card games to extend.
* Added a Player class to represent players in a game.
* Added an Actions enum to represent the "legal action" a player can
take on their turn.
* Added a Roles enum to represent the purpose of a player.
* Updated default exports of enums, models, and types.
* Updated Deck to use JavaScript private fields (TC39 Stage 3). Removed
unnecessary logging statements from the Deck.ts file.
* Installed a babel package so that babel can understand private fields.
* Created a Codenames class to represent a Codenames game. This should
be considered a prototype, though basic testing can be done without a UI
using the notes we've provided. Games (like this one) are for now
expected to be treated as a series of states until a win condition is
met.
* Updated notes with new findings.

Known Concerns:
* There is probably better structuring that can be done for classes,
base classes, and file structure in general. Perhaps some abstractions
are unnecessary. Consider these as you flesh out your first full
prototype.
* It's worth considering adding some kind of "Lobby" class to handle
getting the correct setup of players before starting the game. The game
isn't necessarily responsible for creating the players. That's debatable
though.
* Be wary of how you code in classes. Right now, in the Codenames class,
there are just a few scenarios where you're relying on outside data
under the assumption that it's in sync with your game's data. If things
go normally, that would be the case. But some people could be malicious
and try to manipulate your JS to circumvent regular game rules. Whether
you address that now is up to you.