Get Basic Working Version of UI-Game Interaction
Note that this commit is an early draft mainly intended to get the ball
rolling. There's still a lot of progress and refactoring/cleanup to do.

Main Progress:
--- Frontend
* Added a basic way to play the game through the UI. (Many patches are
still necessary. And a REAL user interface is needed.)
* Add a "CodenamesCard" component to represent a codenames card. It can be
clicked to send "Reveal" actions.
* Added a "Codenames" component to hold the Codenames cards, in addition
to other things necessary for game interaction.
* Added a "GameSelection" component for displaying the list of available
games.
* Added a Login component for players to enter a lobby with a given
name. This should eventually be its own page.
* Added a Vuex store for interacting with the server. It sends actions
for the server to perform.

--- Backend
* Updated the server to respond to actions from the UI. It sends back
actions for the UI to perform.
* Updated "Codenames" class to log game progress on the server side.
Also added an "error" state. This needs to get flushed out.
* Updated the "Game" abstract class to have a method to handle actions.
That way, the "Lobby" class can still perform important actions without
needing to know the exact class of the game.
* Added a method to the "Lobby" class for adding new players.
* Updated the "Player" class to enable it to serialize getters in JSON.
* The Websocket setup is much more robust now.

Extra Progress:
* Updated ESLint rules for the store folder. Vuex encourages mutable
objects.
* See the NOTES for some helpful information.