Refactor/Simplify Deck Creation Logic
* Deleted the concepts of DeckManifests, DeckSchematics, and DeckTypes
since they weren't being used practically and consistently.
* Deleted mappers.ts since it didn't see significant use.
* Changed the Deck class to use simply an array of cards for its
constructor. This shifts reponsibility for Deck creation to the game (or
other entity) seeking to build a deck.
* Updated the Codenames game to create its own deck.
* Updated necessary/impacted tests and files after changes.
* Added a "buildCard" function to the mockBuilders.
  - Added the function to convert an enum to its values (not its keys)
as a utility function in the mockBuilders file.

Note: For common decks that will be created often (like a classic deck
of cards), it may (or may not) be worth considering creating a child
class that uses a special deck layout. If you do this path, try to make
the inheritance simple, and don't create an abstraction too complex
(like in the manifest/schematic/deck-types situation).