Configure Jest for Frontend Tests
* Added a Jest configuration for the frontend.
* Configured Jest to properly handle/transform .vue files.
* Added notes on configuring Jest for testing Vue applications.
* Installed Vue Testing Library, vue-jest, and an extra babel package
for testing Vue components and configuring Jest.

Note: Confirmed that the configuration worked by running the simple
example in the Vue Testing Library docs. The code covergae now properly
shows Vue files as well. If any additional configuration is needed in
the future for Vue quirks, we'll add them as we discover them.